requirements.txt
amqp==2.1.4
billiard==3.5.0.2
celery==4.0.2
defusedxml==0.5.0
Django==1.10.6
django-allauth==0.31.0
django-braces==1.11.0
django-celery-beat==1.0.1
django-celery-results==1.0.1
django-cors-headers==2.0.2
django-crispy-forms==1.6.1
django-debug-toolbar==1.7
django-extra-fields==0.9
django-filter==0.15.3
django-rest-auth==0.9.1
djangorestframework==3.5.4
djangorestframework-camel-case==0.2.0
djangorestframework-filters==0.9.1
djangorestframework-gis==0.11
drf-extensions==0.3.1
drf-nested-routers==0.11.1
gunicorn==19.7.0
kombu==4.0.2
Markdown==2.6.8
oauthlib==2.0.1
psycopg2==2.7
python3-openid==3.1.0
pytz==2016.10
redis==2.10.5
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
sqlparse==0.2.3
vine==1.1.3
python-memcached==1.58
